Fix graph sizing issue; tidy up ComparePage
The graph sizing issue: for the initial render, the size was calculated
as the entire page width (more or less).  After the first render, the
size was calculated again and a second render was performed. This time
the size included the fixed width and paddings of the Container.  This
resulted in the second render having a much narrower width, and the
animation being "odd".